[IMP] various: improves button titles for better command palette
Currently, in some of the important business objects' form views,
few header buttons are missing the title, and so the command palette
displays the button string which is not very clear/useful.

This commit improves the behavior by adding titles to the buttons. Below
are the model wise actions/methods linked with updated buttons:

 - action_rental_quotations_new (crm.lead)
 - action_timer_pause (helpdesk.ticket)
 - action_timer_resume (helpdesk.ticket)
 - action_timer_start (helpdesk.ticket)
 - action_timer_stop (helpdesk.ticket)
 - action_fsm_create_quotation (project.task)
 - action_timer_pause (project.task)
 - action_timer_resume (project.task)
 - action_timer_start (project.task)
 - action_timer_stop (project.task)

Task-2622266

closes odoo/enterprise#20680

Related: odoo/odoo#75935
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>