[FIX] appointment: fix appointment test
JS Part:

During night tests, the triggerPositionalMouseEvent
will not work the way intended because of the dynamic
click position computed. To avoid this, we move to the
next week and hardcode the position for the click.
We also set the time format to avoid issue with
localization format.

Python Part:
Fix tests creating/generting slots of appointment types.
Now we check correctly the slot based on the date provided
and the month of the calendar displayed. In case multiple
months are used (happens in the end of the month) then we
check if the slots of the next month has a slot.

Indeed the slots are only displayed for the month we displayed.
For example: we have a slot for the 1st october.
And we are in september and the first days of october are
diplayed to end the last week of september. Then the slot will be
present only when we displayed the month of october not september.

task-2638269
ENT PR: odoo/enterprise#20692

closes odoo/enterprise#21060

X-original-commit: 0e0ee56002ea13ffe36ebc4bede01a82f8af220d
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>