[FIX] web_enterprise: temp. disable test_not_exist_partner_vcard
When running python test_not_exist_partner_vcard test, the HTTP status
code returned for the tested route isn't always the same (500 vs 404).

This is due to a behaviour difference following the presence (or not) of
the `http_routing` module:
- 500 not installed
- 404 installed

After discussion with the framework, it has been decided to temporarily
disable this test until the current refactoring of the HTTP layer is
finished (and should fix this behaviour difference).

closes odoo/enterprise#20598

Signed-off-by: Xavier Dollé (xdo) <xdo@odoo.com>