[REF] approvals_purchase_stock: adapt test to new incoming picking type
A new picking type for return has been created. The test of this module
doesn't expect more than one picking type of 'incoming' type. This
commit fix the test and ensure to use the classic receipt for approvals
with a purchase and a product storable or consumable

closes odoo/enterprise#20605

Task: 2392001
Related: odoo/odoo#64384
Signed-off-by: Arnold Moyaux <amoyaux@users.noreply.github.com>