[FIX] sign: batch for sign request item
Get correct value for mail template when send sign request in batch.

Before this commit, when send sign requests in batch, the template maynot get correct values.
But since we don't provide any UI to allow user to do this, and no other modele call this in batch,
this bug is not noticeable.

Part-of: odoo/enterprise#19972