[IMP] social_*: revamp social post form view and preview
1. add live_post_link to allow to go to the real post from preview

This commit adds a live_post_link computed field to live_post model.
this field is used in the live post preview and is used to allow the user
to navigate from the post preview to the real post on the target social media.

Note: Instragram live post link is never set for now as the information we have
do not allow us to compute the real link of the post on Instagram.

2. use user timezone to display publish date in post peview

3. add cancel button to post view

This commit adds the possibility to cancel a post if it is scheduled,
in case the user changes his mind in the meantime.

4. add info note about utmized link in the post content.

5. add generic preview method for social.post.template to compute live_post_link
and published_date.

Task-2516278
COM PR: odoo/odoo#76934
ENT PR: odoo/enterprise#21055