[FIX] core: Prevent redundant default on related
Setting a default value on a related field that is readonly without an
inverse method is illegale. We now ensure it at the field instanciation.

If the other related field has a default value then duplicating the
default on the relating field is illegale too. There is no cheap way
to ensure it so we fixed those fields here without including a warning.

For the record, granting a default value to a related readonly=False
field is permitted.

Compagnion of odoo/odoo#67762

closes odoo/enterprise#17313

Signed-off-by: Julien Castiaux <Julien00859@users.noreply.github.com>