[IMP] documents: create a new document request for each recurring upload_file activity
Currently, recurring upload_file activites will be linked to the same document and uploads will be attached as new versions on the existing document.

With this commit, an upload_file `mail.activity` that triggers a future upload_file activity will create a separate document request, with the triggered activity linked to the newly created document record.

Task-2627837

closes odoo/enterprise#21079

X-original-commit: 51a8676aa0b63d79ee481e2beedba7d4a7019c05
Related: odoo/odoo#77004
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>