[FIX] sign: email not being sent if template has no sign items
In task-2475732 a default role was created for templates that had no sign items,
to allow adding new items while signing. However adding this new role prevented
the accesses for the sign request to be sent in templates without sign items. This
PR adds a new check to verify this case and allow again the email to be sent.

X-original-commit: 7abe5fdfddb36ad2780925072f56dd8bae7fbd98
Part-of: odoo/enterprise#20728