[IMP] l10n_mx_edi: allow negative line for mx invoice
CDFI edi does not support line with a negative or null subtotal
but this use case is common when you want to apply a global discount.

This commit modify the cfdi xml generated. The negative line are
splitted on the positive line affecting the discount of the positive
line.

The total of the invoice and the taxes remain unchanged.

closes odoo/enterprise#21995

X-original-commit: 1bf89c5284f941733622c7f4e6b33dbef2b2d678
Signed-off-by: Josse Colpaert <jco@openerp.com>
Signed-off-by: Laurent Smet <las@openerp.com>