[IMP] stock_barcode: change inv. lines buttons
For the Inventory Adjustment in `stock_barcode`, adds a new button to
set/unset the counted quantity (`inventory_quantity` field) on each line
If the line is unset (`inventory_quantity_set`) and the counted qty. is
equal to zero, a '?' will be displayed instead of the qty. to inform the
user this line will be ignored when the inventory adjustment will be
validated.

If this button is clicked, if the line is set, it will set the counted
qty. to 0 and unset the line. If the line is unset, it will set the
counted qty. to the on hand quantity and mark the line as set.
The button will show an orange 'X' if the counted qty. is different than
the on hand quantity. It will show a green 'V' if the counted qty. is
equal than the on hand quantity.

Also, the buttons to increment/decrement the quantity are removed for the inv.
adjustment. Instead, -1/+1 buttons will be shown only for the selected line.

Another change is the button to validate the inventory adjustment was
replaced by the "Apply" button. This button works the same way than the
previous one, except it validate page by page and quit the inventory
adjustment if there is only one location to process.

task-2632818

X-original-commit: 09aa4576467c4c7af9dd637cd92c108298d385f9
Part-of: odoo/enterprise#21810