[IMP] sale_renting_crm: improve the dependency
Currently, 'sale_renting_crm' depends on 'sale_renting' and 'crm',
and adds fields on lead related to rental orders and quotations.
However, it can be installed independently from 'sale_crm', which
can also add fields on lead related to regular orders/quotations.

Problem with this flow is, if both 'sale_crm' and 'sale_renting_crm'
are installed, the rental quotations/orders linked to leads are also
included in the regular ones.

To improve this behavior, we make 'sale_renting_crm' dependent on
'sale_crm', so that we can easily split rental orders/quotations
and rental ones.

Apart from that, this commit does the following improvements:
- renames the stat button string from 'Rental Quotations' to 'Rentals'
  to make it fit in a single line
- re-positions this stat button to push it after 'Quotations' (or
  'Orders' if visible) stat button to make the buttons order consistent
   with header buttons
- removes the default 'Customer' filter while seeing rental orders from
  leads since it is already handled with domain

Note: Few methods are introduced in 'sale_crm' which are overridden
here to make the splitting easy. For more details, see linked
community PR.

taskID-2627580

closes odoo/enterprise#20708

Related: odoo/odoo#83030
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>