[IMP] project_enterprise: ProjectControlPanel for mobile
The reporting views have been converted to OWL so that their extensions
like the project_graph and project_pivot views had to be converted too
along with the control panel they use: ProjectControlPanel.

Here we take care of the appearance of that control panel in mobile mode.

closes odoo-dev/enterprise#211

Related: odoo-dev/odoo#1022
Signed-off-by: Mathieu Duckerts-Antoine <Polymorphe57@users.noreply.github.com>
Part-of: odoo/enterprise#19469