[FIX] industry_fsm: mark as done on the first is_closed stage
Before this commit, when there are two closed stages in a field service
project and the user wants to mark as done a task, he has a traceback
because we give more than one stage in `stage_id` field of this task.

This commit fixes this bug and improve the method calls when the user
marks as done a task.

task-2638359
closes #20700
related PR: odoo/odoo#76166

X-original-commit: 7d85c1cc364d75d196d68d716bc87bad0f0176e9