[FIX] project_enterprise: display project sharing feature
Before this commit, the project sharing is correctly not displayed
because of use the service 'enterprise_legacy_service_provider', this
one need the home menu service. But this service is not used in project
sharing.

This commit removes the using of these services in project sharing
feature.

Related PR: #19483

Part of task-2633229

Community PR: odoo/odoo#76098

closes #20662