[IMP] project_enterprise: prevent date auto shift in the past
Prior to this commit:
When using the date auto shift feature from the task dependencies gantt view,
modifications in the past were allowed. It was also possible to reschedule
a task (through the use of the date auto shift feature) that was in a closed stage.

After this commit:
The user will be warned that the rescheduling could not be done as it
would move the task in the past. Is is also now impossible to reschedule a task
(through the use of the date auto shift feature) that is in a closed stage.

task-2638359
closes #20700
related PR: odoo/odoo#76166

X-original-commit: 931f3501f6dd39f00a5eae40f2209ac2f0caaf60