[FIX] project_enterprise: remove unsed services in project sharing
Before this commit, when web_studio is installed, the user cannot access
to project sharing feature because of 2 services created and adde in the
registry in web_studio: 'studio' and 'studio_legacy' services.
These services need home_menu service but this service is removed in
project sharing assets because we don't use the HomeMenu component and
so, we don't need this service.

This commit create a function to remove these services. This function
will be used before calling the startWebClient function.

Related PR: #19483

Part of task-2633229

Community PR: odoo/odoo#76098

closes #20662

Signed-off-by: LTU-Odoo <IT-Ideas@users.noreply.github.com>