[IMP] stock_barcode: filter on product gs1
In the Barcode App, user can filter pickings by scanning a product's
barcode, but before this commit, that doesn't work with GS1 barcode.
Now, it will try to decompose GS1 barcode and then checks if it has a
product part on it. If it's the case, the pickings will be filtered.

task-2641528

closes odoo/enterprise#25389

Signed-off-by: Arnold Moyaux <arm@odoo.com>