[REF] stock*,mrp*: clean `setUp` vs `setUpClass`
In a lot of test class, we use `setUp` instead of
`setUpClass`. `setUp` is execute for each test method and `setUpClass`
will be execute only once by Class (and use savepoint + rollback).

Then change setUp into setUpClass reduce the time to make all tests
and avoid to repeat this error for the future.

closes odoo/enterprise#21563

Related: odoo/odoo#78082
Signed-off-by: William Henrotin <Whenrow@users.noreply.github.com>