[FIX] project_enterprise: prevent mounting component on sample data or mobile
Prioir to this commit:
The mounting of the ConnectorContainer component and the setup of its re-rendering
was always performed.

After this commit:
The mounting of the ConnectorContainer component and the setup of its re-rendering
will only be done when not in mobile and not when displaying sample data.

task-2638359
closes #20700
related PR: odoo/odoo#76166

X-original-commit: 823fc14db09736a531a19d9ced9536537fb91065