[FIX] project_enterprise: prevent altering data in sample mode in Gantt
Prior to this commit:
As the (+) and magnifier buttons are visible in sample mode in the Gantt view
since this PR #18819, it could be potentially dangerous if used in parallel
with the default_group_by attribute of the gantt. It that case, when the user
uses the buttons, the value of the grouped data would be written together with
the date.

After this commit:
As the (+) and magnifier buttons are not visible anymore, which will prevent
any unwanted behavior.

task-2638359
closes #20700
related PR: odoo/odoo#76166

X-original-commit: d463848f540d05e2b870c3a4ffabdac0ed251e5c