[FIX] pos_settle_due: ask user if the payment is for customer account
Previously, we automatically add the customer account payment when
settling due. This however is not very intuitive so we now implemented
a heuristic such that when the order amounts to zero, it means that
the user is settling due. If no customer account payment in the list
of payments, we ask the user if we add the payment before validating
the order.

X-original-commit: 514e0a1a17d1b64b765f9f3b5df787015deab994
Part-of: odoo/enterprise#20849