[FIX] mrp_plm: show ECOs from all previous BOMs
The ECO status button on BOM form will lead you to the ECOs link to this
BOM. Since every eco creates a new bom, the ECOs link to the old BOMs
won't be shown after you click the button. In this commit, we also count
and show ECOs link to previous BOMs.

Task-2639200

X-original-commit: d2b97b6d8cb7e810caf5bf081dcb3ed965d133b5
Part-of: odoo/enterprise#21531