[FIX] stock_barcode_picking_batch: don't autobatch
Before this commit, when the user tries to open a picking in the Barcode
App, Odoo will check if this picking can be batched with other pickings.
With this commit, it will no longer ask if the picking the user wants to
open has no partner.

Task: fp-request
X-original-commit: 109504ed1a3db7970345d9077c2486025442ab32
Part-of: odoo/enterprise#21851