[REF] project_enterprise: refactor task auto-shift
This commit refactors the "auto-shift" functionnality introduced in PR
 #18045.

Purpose
=======

The auto-shift functionnality shifts the tasks backward or forward
thanks to two triggers :
- Manual trigger : User clicks on an arrow in the gantt view
- Auto trigger : User plans a task in conflict with a depending task
that should be shifted in order to keep a waterfall order.

This functionnality may be summarized as :
1) When a forward (resp. backward) shift is triggered, find the first
datetime in the future (resp. the past) where the assignee is available.
2) Starting from this datetime, allocate the planned hours based on
user's calendar availabilities.

Analysis
========

Currently the implementation takes a big advantage of caching the work
intervals information for a given calendar and resource and fill
efficiently the dict in order to limit the number of calls to
work_intervals_batch.

However efficient, the current implementation can be factorized in order
to have :
1) a unique method responsible of the work intervals caching which is
handled two times currently.
2) a unique method responsible of giving the new begin and end planned
dates, whatever there are or not planned hours on task
3) an efficient way to use the 'calendar validity' API of resource
introduced in PR #21135. Which allows to remove overrides in
project_enterprise_hr and project_enterprise_hr_contract.

Specification
=============

The solution is based on a new cache definition : The work intervals
cache caches work intervals for a given resource entity; a resource
entity is either a resource.resource, either a res.company (this is done
in order to cache the default work intervals of the company when a user
isn't available for a given period (not under contract for example)).
This cache is used in combination of a dict `resource_validity` which
gives the intervals where the resource's calendar must be queried
('valid' intervals) and the other, where it's the company's calendar
that should be used ('invalid' intervals). The sum of those two ('valid'
and 'invalid') intervals is actually the whole interval queried for a
given resource. For any time in this intervals, the data are already
cached.

The problematic spotted in the analysis are adressed this way :
1) The `_update_intervals_cache` method takes all the responsibilities
regarding the cache updates.
It first check that the interval has not already been queried for the
given resource. If not, it searches the calendars of the resource and
its work intervals.
If the resource is not valid (for example not under contract), it
updates and searches into the company's work intervals cache.
This method is used to search for the first possible datetime to plan
the shifted task and then to plan the planned hours of the task.
2) The `_get_planned_dates_auto_shift` cope with the problematic of
dividing code into two parts relative to the fact that the task has
planned hours of not. It first gets the first possible datetime to plan
the task, then apply the current duration for tasks without planned
hours or search for work intervals to plan the current numbers of
planned hours on the task, for tasks with planned hours.
3) The `_get_resource_calendars_validity` takes all the advantages
of the resource API to retrieve calendars and return the user validity.
The fact that hr or hr contract is installed is thus transparent.

Functionnal change log
======================
- Fallback calendar is always the company calendar whenever the employee
is invalid (not under contract, not yet created, already fired, ...)
- Between two contracts, there is always a fallback on the company
calendar, regardless of the length of the period between the two
contracts.
- This commit changes tests in order to take into account the few
differences introduced during the refactor, i.e. for employees of type
employee or student, the fallback calendar is ALWAYS the company
calendar and never the resource's calendar.

task-2643752

closes odoo/enterprise#20768

Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>