[FIX] timesheet_grid: fix error when user create their own timesheet
Currently, when we connect as a timesheet user with the timesheet module
and try to create a new timesheet entry, it throws an error because
timesheet user does not have right to update the 'validated' field and
due to inverse method of 'validated_status' it was going to update that field.

So in this commit, remove the inverse method of the 'validated_status' as
it was only used in statusbar to display the non-clickable statusbar. so
no need to add the inverse method.

closes odoo/enterprise#21222

Taskid: 2629541
X-original-commit: dee80deae540ae713a11d51888750adcab180f11
Signed-off-by: LTU-Odoo <IT-Ideas@users.noreply.github.com>