[FIX] stock_barcode_picking_batch: autobatch doesn't care about the date
When user wants to open a picking which is not a part of a batch, Odoo
will ask if they want to group it into a batch if it find other pickings
to batch with.
Before this commit, to find those pickings, Odoo searched for pickings
who should be done before the selected one or before today.
Now, it will no more take the date for the search.

Task: fp-request
X-original-commit: 3907e581c55964f0223c06117f769cf734e2a98b
Part-of: odoo/enterprise#21851