[FIX] project_enterprise: prevent traceback in gantt
Before this commit:
When having the gantt grouped by user_ids, without any data, a traceback
was raised when creating a new record through the use of the (+) button.

After this commit:
No more traceback is raised when creating a new record using the (+)
button even when no data is displayed in the grid.

Issue introduced in #19780 when changing from use_id to user_ids.

task-2638359
closes #20700
related PR: odoo/odoo#76166

X-original-commit: d87ce8a0d22f06aebf2014351145d34a44f103ba