[FIX] web_cohort: click on aggregate in board
Let us assume we have a dashboard view that has a cohort subview and an
aggregate with measure attribute set to "__count__" like in
crm_opportunity_view_dashboard.
Add that view to the board application, reload, and go to board app.
Click on the aggregate. A crash occurs because a legacy cohort view will
be instantiated with cohort_measure="__count__" in its context.

Here we make the legacy cohort view process its measure to make sure that
it does not use "__count__" anymore.

closes odoo/enterprise#20778

X-original-commit: b8a217892a56431f8587240b1cc8960a394d8315
Related: odoo/odoo#76368
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>
Signed-off-by: Mathieu Duckerts-Antoine <Polymorphe57@users.noreply.github.com>