[IMP] social_demo: complete social demo data and fake upload button
This commits reviews the demo data for social in order to be more complete
and to cover more of the cases it's supposed to demonstrate.

1. Display a toaster when uploading video for youtube telling the user them
cannot upload the video in demo mode.

2. revamp youtube demo data -> use a real video in order to get the thumbnails
and embed video instead of placeholders of non existing video from youtube.

3. complete social demo data to be able to compute stream post link

Task-2516278
COM PR: odoo/odoo#76934
ENT PR: odoo/enterprise#21055