[FIX] web_enterprise, web_dashboard: prevent subviews's control panel to scroll
In web_enterprise, the control panel adapts its position according to the scroll.
In web_dashboard, we don't want to enable this feature for the subviews.

So, this commit introduces a key "adaptToScroll" in the "display" props.
If absent or truthy, the control panel will adapt to the scroll.
If present and falsy, it won't.

closes odoo/enterprise#20796

X-original-commit: 6c91e3c837ea6aa3ec9b95b897f0ff6d01052e8f
Related: odoo/odoo#76390
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>