[IMP] documents: allow to resize searchpanel
PURPOSE

If you work with long workspace names, they can be hidden by the size of the
left menu. In that case, we don't have any option to see the full name.

SPECIFICATIONS

Allow the user to resize the search panel and Display the name of the
workspace on hover

Task-2641699

closes odoo/enterprise#20917

Signed-off-by: William Braeckman (wbr) <wbr@odoo.com>