[FIX] planning: fix traceback on empty date of planning slot
Currently when only planning is installed and if we remove the
start date then it will generate an traceback due to add days in bool.

So in this commit fix the issue by checking date value exist then only
add days or compare else set it as Fasle.

closes odoo/enterprise#21072

Taskid: 2642872
Closes: #20763
X-original-commit: fc4c782b6ee08a7219972058a79a9e6f7a39478f
Related: odoo/odoo#76967
Signed-off-by: LTU-Odoo <IT-Ideas@users.noreply.github.com>