[FIX] helpdesk_fsm: allow portal user to click on ticket button
Before this commit, in project sharing feature, when the portal user is
in task form view, he can see the ticket stat button if a ticket is
defined in the task but this button is disabled.

This commit allow the portal user to click on this button if he has
access to the ticket. If he cannot then the stat button will be
invisible since it brings no more information. If he has access and
clicks on this button, then he will be redirected to `/my/ticket/<id of
the ticket>`.

task-2648955

Community PR: odoo/odoo#82379

Closes #23315