[FIX] helpdesk_sale_timesheet: select any billable project
Before this commit, we keep the previous behavior for the selection a
project in a helpdesk team. That is, only a project with a
pricing_type='task_rate' can be selected in the helpdesk team.
Before, the problem was because of a task linked to a helpdesk ticket.

Since it is not the case now, this commit removes the condition on the
pricing_type, now the user can select a billable project with
allow_timesheet enabled.

task-2638359

Community PR: odoo/odoo#76166

closes #20700

X-original-commit: 3192909932af2321cdb71e9b14b537065b8ac561