[REF] web_dashboard: call useSubEnv on each sub view
Before this commit, the DashboardView used the hook "useSubEnv" to
generate a particular env for its sub views. The purpose of this
env was to be able to extract the local and global states of those
sub views and their context, by overriding some internal keys used
for the communication between the ActionService and the view.

However, the "useSubEnv" hook also changes the env of the component
using it. As a consequence, doing so implies that we can't call the
"useSetupView" hook where we want (it must be called before calling
"useSubEnv", hence the comment in code).

This commit defines an higher order component to use around sub
views, and calls the "useSubEnv" in that component. That way,
the env of the dashboard view is kept untouched and each sub view
has its own internal keys to communicate with the dashboard. It
thus also simplifies the code that extract information from the sub
views, as it doesn't need to try to guess which information came
from which sub view anymore.

X-original-commit: 8d110464eb8cace2b488bcab371bd9d1c6e39c99
Part-of: odoo/enterprise#20894