[FIX] industry_fsm: display the planned date begin in project sharing
Before this commit, we display the planned date begin in each kanban
card of task like it is done in Field Service. But the problem is in
Field Service, we use a context according to the action called. This
context is not used in project sharing and so the date is not correctly
formatted.

This commit removes the using of this contexte in project sharing
feature and display the planned date begin field instead of format and
display it with momentjs.

task-2633229

Community PR: odoo/odoo#/77156
closes #21155

X-original-commit: a7c20684592e60ec0608dbf81461b51ba1d616d6