[IMP] mrp_workorder_plm: imporve operation change tracking
Track the quality points change of every operation by using their title.
If the title of the points on the old and new BOM are the same, we
consider any change an "update". If one title shows in new BOM but not
in old one, we consider it's "add". If one title shows in the old BOM
but not in the new one, we consider it's "remove".
We assume there won't be duplicate titles for quality points of the same
operation.

Task-2639200

X-original-commit: 0244b6fcaf0ecbb33f188a594ccc732be1f2ec59
Part-of: odoo/enterprise#21531