[FIX] industry_fsm: hide the phone and address of the partner
Before this commit, in project sharing, the portal can see the phone and
the address of the customer if the task is a fsm task. The problem is if
the portal user is not the customer, we don't respect the privacy
between our company and our customer. Perhaps, this customer does not
want to share its address to every external collaborators.

This commit hides those fields and will be removed in master.

task-2633229

Community PR: odoo/odoo#/77156
closes #21155

X-original-commit: f90cd5434015fd0bdcd63c0e8e7a0406ca6dfdd9