[FIX] industry_fsm_sale: display the material button if material > 0
Before this commit, the portal user can see the Material Products stat
button even if no material is added to the task. Since this button is
disabled in project sharing, it is not useful to show this button when
no material is added to the task.

This commit adds a condition for the visibility of this button. This
button will be visible only if:

 1) the customer is set;
 2) allow_material=True in the project shared;
 3) the number of materials is greater than 0.

task-2633229

Community PR: odoo/odoo#/77156
closes #21155

X-original-commit: 96d1f9452456e735ec60c5a90a14443c3f211b05