[FIX] planning: fix employee avatar with no slot
Prior to this commit, the employee avatar wasn't shown if there wasn't
any slot scheduled for the period displayed in the gantt.

With this commit, the id of the resource is returned with the planning
hours info, in order to know which employee is linked to the resource.

Linked PR : #19695 and #19647

How to reproduce the issue ?
- Install Planning app
- Click on Planning
- Go to a week for which an employee has no scheduled slot
- The employee has no avatar neither its number of planned hours/working
hours

How the issue was fixed ?
The employee_id of a resource cannot be accessed directly from the
resource id. As the server is queried in order to display the work
information, the call returns also the employee id. This employee id is
used to display the avatar employee widget, if set.

PR : #20835

task-2645575

X-original-commit: 26399b31acc75a18c8b4d2d01b4fac5b2f3f0ac6
Part-of: odoo/enterprise#20881