[FIX] stock_barcode: use two times same lot
Before this commit, if the user wants to scan the same lot for multiple
lines for the same product (in a batch pickings for example), it's not
allowed because the application check the serial number isn't already
used (as a SN has to be unique) but it do the same check for lot, which
shouldn't happened. Also, it should raise an error but it didn't.

How to reproduce:
 - Create a product tracked by lot;
 - Create two different receipts for this product and batch them;
 - Confirm the batch and open it in the Barcode App;
 - Scan the product -> A line is selected;
 - Scan the lot -> The lot name should be added to the line and its
   quantity should be incremented.
   Repeat until the line is fulfilled.
 - Once the first line is fulfilled, scan again the same lot
   -> The second (unfulfilled) line isn't selected or incremented, the
      first one isn't incremented neither.

The wanted behavior is, when a lot is scanned and the current line has a
product tracked by lots:
 - Increment it if the line isn't fulfilled.
 - Try to find an another unfulfilled line for this product if the
   current line is fulfilled, and increment it.
 - Increment the current line even if it's fulfilled if there is no
   other unfulfilled line for the same product.

task-2646107

X-original-commit: 7b3b948e80eb4b143bb4f2cb4bc8c64792d154cd
Part-of: odoo/enterprise#21851