[FIX] sale_planning: sanitize planning context
This commit fixes an issue when clicking on a task on the Plan dialog of
the planning gantt view.
This issue was due to the fact the context value was undefined rather
than null.

linked PR : #19647

How to reproduce the issue ?
- Install Project Forecast and Sales Management;
- Create a plannable product which creates Project and Task on order;
- Create a Sale Order with this product and confirm it;
- Go to the Planning app
- Click on the magnifying glass
- Click on the Task link
- Traceback due to undefined context value

How to fix it ?
- Sanitize the context.

PR: #20835

task-2645575

X-original-commit: 6b804998f516b559ab5d991f0d1d37cbfdee52e7
Part-of: odoo/enterprise#20881