[FIX] web_dashboard: get the group_operator from the field description
Before this commit, a group_operator set on a field in python was never considered by the dashboard view.

After this commit, when a group operator has been defined on a field in python,
it is used if no group_operator has been specified for that field in the dashboard view arch.

closes odoo/enterprise#20866

X-original-commit: 26cbaabe8a25a5fc558508700d6c498fbe4cc099
Signed-off-by: Mathieu Duckerts-Antoine <Polymorphe57@users.noreply.github.com>