[FIX] stock_barcode: increment selected line
Before this commit, when a product or a lot is scanned and all lines
about this product or this lot are fulfilled, it's always the same line
who will be incremented (the one the application will find at first).
So, if the user manually selects a line in order to increment
specifically this one, the application doesn't care.

How to reproduce (need batch picking enabled):
  - Create a product tracked by lots;
  - Create two receipts for this product and batch them;
  - Confirm the batch and open it in the Barcode App;
  - Scan the product then scan a lot until the line is fulfilled;
  - Then keep scanning the same lot -> It should select the other
    picking's line for this product, add the lot name on it and
    increment its quantity done.
  - Scan the same lot until all lines for this product are fulfilled;
  - From here, if you scan again the lot, no matter which line you
    selected, it will always increment the same line.
    -> It should increment the selected line instead.

task-2646107

X-original-commit: cc539f0e6e8a82e02abe26e1d2ad4e3285244938
Part-of: odoo/enterprise#21851