[FIX] {sale_,}planning: fix demo data use case
This commit fixes the demo data use case where there were a mix between
a restaurant planning and a service provider company.

The HR demo data suggest that the standard use case is a service
provider company - and specifically in the software development field.

The planning demo data are adapted to suit this use case.

PR : #20883

task-2645575

closes odoo/enterprise#20907

X-original-commit: 360033c90c06897135b073031ee6f50f1402413a
Signed-off-by: LTU-Odoo <IT-Ideas@users.noreply.github.com>