[FIX] web: fix concurrency issue in cohort
Let's assume the following scenario in a cohort view
 - toggle a filter on a slow network s.t. it takes a while to load
 - while it's loading, toggle a measure

Before this commit, the view was kind of locked after those two
steps, i.e. toggling measures did not work anymore (it could be
unlocked by interacting again with the search view though). The
reason is that the new filter triggered a re-rendering of the
withSearch component (the parent of the cohort view), thus calling
willUpdateProps on the cohort view, which reloads the data. Toggling
a measure cancels the current load operation (via keepLast), and the
promise returned by willUpdateProps is thus pending forever. From
that point, only another re-rendering from the parent (withSearch)
could unlock the situation by cancelling the previous rendering.

To solve this, we use the new Race primitive. It ensures that the
promise returned by load resolves as soon as the view is actually
reloaded (whether it came from the load itself, of by toggling a
measure).

closes odoo/enterprise#20992

X-original-commit: 8ea8c0121811d0446e1c435e2e8bdc4d20e1f991
Related: odoo/odoo#76778
Signed-off-by: Géry Debongnie (ged) <ged@openerp.com>