[FIX] web_enterprise: correctly catch error when scanning a barcode
Since 9f5f38c67ff59b11d99000d7951f7a2703402f97 , we refactored the
BarcodeScanner to use OWL. This new implementation didn't manage error
when scanning (e.g. if we refuse the permission to access the camera)
an an error is raised.

Now, the "BarcodeDialog-Error" event is correctly triggered and the
error is catched in the "barcode mobile mixin".

Add the translations file for the "barecode_mobile" module.

Task-2647029

closes odoo/enterprise#21137

X-original-commit: 42036f534a199caa1f75743c2df6cd35abf99630
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>