[FIX] web_enterprise: show filters on account reports on mobile
In some custom action/view, the control panel might have
"searchViewButtons" with no search bar.

But on mobile, the 'searchViewButtons' is only shown in the separated
mobile search view and users cannot then click on the search bar
to open this view.

To fix this issue, we add the 'searchViewButtons' on the main component
of the control panel even if there is no search bar.
Note that "web_enterprise.ControlPanel" is only rendered on small screens.

Steps to reproduce:
* Open Odoo on Mobile
* Go to "Accounting"
* Open the "Burger Menu"
* Select "Reporting" -> "Partner Reports" -> "Partner Ledger"
* BUG => there is no "searchViewButtons" in the Control Panel

opw-2637322

closes odoo/enterprise#20963

X-original-commit: 3852242a88831b339b0e47d2a5fc49f4a66ebf55
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>
Signed-off-by: rfr-odoo <rfr-odoo@users.noreply.github.com>