[IMP] stock_barcode: don't move inventory line
In the Barcode App, a completed line is moved to the bottom of the lines
list to have only remaining lines to process on the top.
But with the new way to process an inventory adjustment, this behavior
isn't convenient anymore.
So, it will not be applied anymore to the inventory line.

X-original-commit: a731463b141b227001d0633748629cf448b7e294
Part-of: odoo/enterprise#21810