[FIX] planning: fix multi-company issue when unassigning
The self.env.company is not reliable when it's called from a recompute.
When the resource is updated to False, the compute is called from the
flush method, and the self.env.company has the value of the default
company of the user.

Keep the company_id as already defined if there are no resource.

This fixes the regression introduced in : #19695

PR : #20934

task-2642872

X-original-commit: 8aaf37fec3f09a4238b4e5bc7fb34e4b86926fe4
Part-of: odoo/enterprise#20959