[REF] web_dashboard: alternative to ControlPanelBottom
With this commit, we stop using the ControlPanel BottomContent
feature, that is about to be removed, in the dashboard. It was
used to add a switch button in the control panel of each sub view.

This commit uses the ViewWrapper to do this. It is no longer an
higher order component for sub views, but a component that wraps
the subview into a div, alongside the switch button, which is then
correctly positionned in the top right corner of the subview.

This commit also ensures that the new dashboard view doesn't use
the style of the legacy one.

X-original-commit: 43f1c02101b1ede88d0e84c26ceb52b02dd1cc0c
Part-of: odoo/enterprise#20990