[FIX] account_disallowed_expenses{_fleet}: fix unfold and indentation bug
Currently, there is a bug that occurs when trying to unfold lines in the report.
They just disappear.

There is also an indentation issue that makes the report unclear.

And finally, when creating disallowed expenses categories,
it is possible to create and edit a new company.
That behavior is not needed there.

closes odoo/enterprise#21911

X-original-commit: 5fadbfcedad215b6622e472f313a749390c2baa6
Signed-off-by: William André (wan) <wan@odoo.com>