[FIX] account_reports: don't display the fiscal position selector on every report
It should only be seen on tax reports, and was available everywhere. This was due to a wrong computation of available_vat_fiscal_positions in case the fiscal position country of the report was None. This available_vat_fiscal_positions options should only be used to contain the fiscal positions that are available for selection, not the ones that are used in the computation (so: it's possible to take all fiscal positions into account without having any set in available_vat_fiscal_positions).

closes odoo/enterprise#21082

X-original-commit: a2eba97781c4ca99e6f0d63f60a05dce402cb3e2
Signed-off-by: Laurent Smet <smetl@users.noreply.github.com>
Signed-off-by: William André (wan) <wan@odoo.com>