[FIX]account_reports:analytic display archived acc
Issue adressed:
This commit aims to display in the analytic report the analytic accounts
that are archived but used on the date range selected for the report.

Archiving an account means we shouldn't be able to use this account
anymore on new account_move. Not that it shouldn't be displayed in the
analytical report.

Because an example is worth a thousands explanation:

let's say we have this:
| account   | balance  | active  | group   |
| :-------- | -------: | :------ | :------ |
| account_a | amount_a | False   | group_x |
| account_b | amount_b | True    | group_x |

Before this commit:

display on the analytical report:

| name      | balance             |
| :-------- | ------------------: |
| group_x   | amount_a + amount_b |
| account_b | amount_b            |

In this situation, first we can't see the account_b even if it was used
in the current date range.
Second, group_a displays an inconsistant amount considering the displayed
information about the accounts.

After this commit:

display on the analytical report:

| name      | balance             |
| :-------- | ------------------: |
| group_a   | amount_a + amount_b |
| account_a | amount_a            |
| account_b | amount_b            |

account_b is displayed if it has been used at least once in the report
date range.

closes opw-2585495

closes odoo/enterprise#21267

X-original-commit: 32beada5d4d3996a961d53b1bd4dec389f855bfb
Signed-off-by: Florian Gilbert <FlorianGilbert@users.noreply.github.com>
Signed-off-by: Brice Bartoletti <Levizar@users.noreply.github.com>